Intialize user management.

+++ routes/web.php
@@ -17,16 +17,21 @@ Route::get('/', ['as'=>'home',function () {
 
 Auth::routes();
 
+
+
+//Report route and actions which is allowed to admin only
 Route::group(['prefix'=>'report', 'middleware' => 'auth'], function(){
-    Route::get('/data', ['as'=>'report.list', 'uses'=>'ReportController@listData']);
     Route::get('/', ['as'=>'report.index', 'uses'=>'ReportController@index']);
-    Route::get('create/', ['as'=>'report.create.step1', 'uses'=>'ReportController@create']);
-    Route::get('create/{id}', ['as'=>'report.create.step2', 'uses'=>'ReportController@createStepTwo']);
-    Route::post('create', ['as'=>'report.store', 'uses'=>'ReportController@store']);
+    Route::get('/data', ['as'=>'report.list', 'uses'=>'ReportController@listData']);
+    Route::get('create/', ['as'=>'report.create.step1', 'uses'=>'ReportController@create'])->middleware('role:admin');
+    Route::get('create/{id}', ['as'=>'report.create.step2', 'uses'=>'ReportController@createStepTwo'])->middleware('role:admin');
+    Route::post('create', ['as'=>'report.store', 'uses'=>'ReportController@store'])->middleware('role:admin');
     Route::get('{id}', ['as'=>'report.show', 'uses'=>'ReportController@show']);
-    Route::get('{id}/edit', ['as'=>'report.edit', 'uses'=>'ReportController@edit']);
-    Route::put('{id}', ['as'=>'report.update', 'uses'=>'ReportController@update']);
-    Route::delete('{id}', ['as'=>'report.destroy', 'uses'=>'ReportController@destroy']);
+    Route::get('participate/{id}', ['as'=>'report.getParticipate', 'uses'=>'ReportController@getParticipate']);
+    Route::put('participate/{id}', ['as'=>'report.putParticipate', 'uses'=>'ReportController@putParticipate']);
+    Route::get('{id}/edit', ['as'=>'report.edit', 'uses'=>'ReportController@edit'])->middleware('role:admin');
+    Route::put('{id}', ['as'=>'report.update', 'uses'=>'ReportController@update'])->middleware('role:admin');
+    Route::delete('{id}', ['as'=>'report.destroy', 'uses'=>'ReportController@destroy'])->middleware('role:admin');
 });
 
 Route::group(['prefix' => 'defect'], function () {
